# src/openapi.yaml
openapi: 3.1.0
info:
  title: Time-addressable Media Store
  description: 'The Time-addressable Media Store (TAMS) is used for storing segmented
    media Flows.

    The TAMS consists of one or more Storage Backends for the media Flow Segment Objects
    and a service providing a database index of the Flow Segments.

    This document is a specification of the service API.


    See the [bbc/tams](https://github.com/bbc/tams) repository for more background
    on TAMS.


    **Note**: the examples provided in this specification are for a system which uses
    a Storage Backend that provides HTTP URLs for uploading and downloading Media
    Objects in buckets.

    This could for example be implemented using an AWS S3 compatible store using presigned
    URLs or by a simple file system storage with an HTTP frontend.

    Clients should parse the [/service](#/operations/GET_service) endpoint and handle
    the Storage Backend type in use appropriately.

    '
  version: '8.0'
  contact:
    name: BBC R&D - Cloud-Fit Production Team
    email: cloudfit-opensource@rd.bbc.co.uk
    url: https://github.com/bbc/tams
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: http://localhost:4010
  description: Local mock of API
- url: https://example.com/tams/{version}
  description: Example TAMS service
  variables:
    version:
      description: API version
      default: v8.0
security:
- bearer_token_auth: []
- basic_auth: []
paths:
...
  /objects/{objectId}:
    head:
      summary: Media Object Information
      description: Return Flow references and other information about Media Objects.
      operationId: HEAD_objects
      tags:
      - Objects
      parameters:
      - name: objectId
        in: path
        description: The Media Object identifier. The Object ID may include special
          characters such as `/` which should be URL encoded.
        required: true
        schema:
          type: string
      - name: verbose_storage
        in: query
        description: 'Include storage metadata in `get_urls`.

          When `verbose_storage` is `false` only `url`, `presigned`, and `label` will
          be included in `get_urls`.

          '
        schema:
          default: false
          type: boolean
      - name: accept_get_urls
        in: query
        description: 'A comma separated list of labels of Media Object `get_urls`
          to include in the response.

          Omitting `accept_get_urls` will result in no filtering of `get_urls`.

          An empty `accept_get_urls` results in an empty or no `get_urls` in the response.

          Media Object `get_urls` with no label will only be returned if `accept_get_urls`
          is omitted.

          Without `get_urls`, the response from the service could be substantially
          faster if it is not required to

          generate a large number of pre-signed URLs for example.

          Where multiple filter query parameters are provided, the returned `get_urls`
          will match all filters.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-label-list.json
      - name: accept_storage_ids
        in: query
        description: 'A comma separated list of `storage_id`s of Media Object `get_urls`
          to include in the response.

          Omitting `accept_storage_ids`, or providing an empty `accept_storage_ids`
          will result in no filtering of `get_urls`.

          Media Object `get_urls` with no storage ID will only be returned if `accept_storage_ids`
          is omitted or empty.

          A full list of available `storage_id`s may be found at the `service/storage-backends`
          endpoint.

          Where multiple filter query parameters are provided, the returned `get_urls`
          will match all filters.

          '
        schema:
          type: string
          pattern: ^([0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12})(,[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12})*$
      - name: presigned
        in: query
        description: 'If set to `true`, only presigned URLs (i.e. those whos `presigned`
          property is `true`) will be returned in `get_urls`.

          If set to `false`, only non-presigned URLs (i.e. those whos `presigned`
          property is `false`) will be returned in `get_urls`.

          If omitted, both presigned and non-presigned URLs will be returned.

          If `presigned` is set to `false`, the response from the service could be
          substantially faster if it is not required to

          generate a large number of pre-signed URLs.

          Where multiple filter query parameters are provided, the returned `get_urls`
          will match all filters.

          '
        schema:
          type: boolean
      - name: flow_tag.{name}
        in: query
        description: 'Filter `referenced_by_flows` on tag values. This option is the
          same as the `tag.{name}` query parameter on the `/flows/` API endpoint.

          '
        schema:
          type: string
      - name: flow_tag_exists.{name}
        in: query
        description: 'Filter `referenced_by_flows` on tag names. This option is the
          same as the `tag_exists.{name}` query parameter on the `/flows/` API endpoint.

          '
        schema:
          type: boolean
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the service implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/trait_resource_info_head_400'
        '404':
          description: The requested Media Object does not exist. 404 MUST be returned
            if the ID has been assigned via the [`/flows/{flowId}/storage`](#/operations/POST_flows-flowId-storage),
            but not yet registered against a Flow Segment.
    get:
      summary: Media Object Information
      description: 'Contains Flows that references the Media Object and other information.


        The paging query parameters and headers are required for the list of Flow
        references in the Media Object.

        Service implementations should return a complete list of Flow references within
        reason and API clients should expect paging to happen in some rare cases where
        a Media Object is used in many Flows.

        '
      operationId: GET_objects
      tags:
      - Objects
      parameters:
      - name: objectId
        in: path
        description: The Media Object identifier. The Object ID may include special
          characters such as `/` which should be URL encoded.
        required: true
        schema:
          type: string
      - name: verbose_storage
        in: query
        description: 'Include storage metadata in `get_urls`.

          When `verbose_storage` is `false` only `url`, `presigned`, and `label` will
          be included in `get_urls`.

          '
        schema:
          default: false
          type: boolean
      - name: accept_get_urls
        in: query
        description: 'A comma separated list of labels of Media Object `get_urls`
          to include in the response.

          Omitting `accept_get_urls` will result in no filtering of `get_urls`.

          An empty `accept_get_urls` results in an empty or no `get_urls` in the response.

          Media Object `get_urls` with no label will only be returned if `accept_get_urls`
          is omitted.

          Without `get_urls`, the response from the service could be substantially
          faster if it is not required to

          generate a large number of pre-signed URLs for example.

          Where multiple filter query parameters are provided, the returned `get_urls`
          will match all filters.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-label-list.json
      - name: accept_storage_ids
        in: query
        description: 'A comma separated list of `storage_id`s of Media Object `get_urls`
          to include in the response.

          Omitting `accept_storage_ids`, or providing an empty `accept_storage_ids`
          will result in no filtering of `get_urls`.

          Media Object `get_urls` with no storage ID will only be returned if `accept_storage_ids`
          is omitted or empty.

          A full list of available `storage_id`s may be found at the `service/storage-backends`
          endpoint.

          Where multiple filter query parameters are provided, the returned `get_urls`
          will match all filters.

          '
        schema:
          type: string
          pattern: ^([0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12})(,[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12})*$
      - name: presigned
        in: query
        description: 'If set to `true`, only presigned URLs (i.e. those whos `presigned`
          property is `true`) will be returned in `get_urls`.

          If set to `false`, only non-presigned URLs (i.e. those whos `presigned`
          property is `false`) will be returned in `get_urls`.

          If omitted, both presigned and non-presigned URLs will be returned.

          If `presigned` is set to `false`, the response from the service could be
          substantially faster if it is not required to

          generate a large number of pre-signed URLs.

          Where multiple filter query parameters are provided, the returned `get_urls`
          will match all filters.

          '
        schema:
          type: boolean
      - name: flow_tag.{name}
        in: query
        description: 'Filter `referenced_by_flows` on tag values. This option is the
          same as the `tag.{name}` query parameter on the `/flows/` API endpoint.

          '
        schema:
          type: string
      - name: flow_tag_exists.{name}
        in: query
        description: 'Filter `referenced_by_flows` on tag names. This option is the
          same as the `tag_exists.{name}` query parameter on the `/flows/` API endpoint.

          '
        schema:
          type: boolean
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the service implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/objects-get-200.json
              schema:
                $ref: vendor/bbc-tams/api/schemas/object.json
        '400':
          description: Bad request. Invalid query options.
        '404':
          description: The requested Media Object does not exist. 404 MUST be returned
            if the ID has been assigned via the [`/flows/{flowId}/storage`](#/operations/POST_flows-flowId-storage),
            but not yet registered against a Flow Segment.
...
components:
  responses:
...
    trait_resource_info_head_400:
      description: Bad request. Query parameters are invalid.
...
  parameters:
    trait_paged_limit:
      name: limit
      in: query
      description: Restrict the response to the specified number of results. Service
        implementations may specify their own default and maximum for the limit
      schema:
        type: integer
    trait_resource_paged_key:
      name: page
      in: query
      description: Opaque string used by backend to access a specific page of results.
        Clients should read the next URL from the `Link` header returned with responses,
        or use value of the returned X-Paging-NextKey header. If not supplied, the
        first page is accessed. Service implementations should ensure a consistent
        sort order is applied to pages of results.
      schema:
        type: string
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
      description: 'HTTP basic authentication using a username and password.


        This option should only be used in constrained cases where Bearer tokens are
        impractical, since the password is intrinsically long-lived, and managing
        user credentials across systems is difficult compared to using centralised
        auth (e.g. via OAuth2) with Bearer tokens

        '
    bearer_token_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'HTTP bearer token authentication.


        The method through which a client should acquire a bearer token is assumed
        to be an OAuth2 grant (either the Authorization Code or Client Credentials
        grant), however the OpenAPI `OAuth Flow` object cannot be used because it
        requires that the relevant URLs are known in advance.


        Clients should provide options to allow one (or both) grant types to be configured,
        as appropriate.


        Where tokens are not JWTs, or are acquired through some other mechanism, the
        API key method should be used instead.

        '
...